Added tentative middleware handling to the framework.
Framework users create a list of middleware packages in their Project
Registry, which are parsed into the application when it runs and applied
at different stages in the request lifecycle.

A middleware package should expose a public interface that will create a
struct of type Middleware for use in the user's application.  Middleware
currently implements two methods - ProcessRequest, and ProcessResponse.

The middleware handler in http.baseHandler will iterate over the list of
middleware processors and execute ProcessRequest, then execute the base
handler from the URLConf, and finally iterate backwards over the list of
middleware processors executing ProcessResponse.

The serverHeaderMiddleware has been created to provide an example of
successful middleware implementation.

Signed-off-by: David Arbuckle <dintlu@gmail.com>